requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
anthropic==0.28.0
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
